requirements.txt
PySide6
psutil
numpy
SQLAlchemy
psycopg2-binary
python-dotenv
alembic
